# loans/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Loans microservices REST API documentation
  description: BaopenBank Loans microservice REST API Documentation
  version: 1.0.0
  contact:
    name: Bao Tran
    email: baopen@gmail.com
    url: https://baopenbank.com
  license:
    name: Apache 2.0
    url: https://baopenbank.com

paths:
  /api/loans:
    get:
      summary: Retrieve Loans REST API
      description: REST API to retrieve new Loan & Customer inside Baopenbank
      parameters:
        - name: mobileNumber
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: HTTP Status 200
        404:
          description: HTTP Status 400 - Not Found Loans with given input mobileNumber


    post:
      summary: Create Loans Account REST API
      description: REST API to create a new Loan Account inside Baopenbank
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Loan"
      responses:
        "201":
          description: Loan account created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "400":
          description: Bad request due to validation errors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"


    put:
      summary: Update Loans Account REST API
      description: REST API to update new Loan Account inside Baopenbank
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Loan"
      responses:
        "200":
          description: Loan account updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "417":
          description: Exception failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"



    delete:
      summary: Delete Loans REST API
      description: REST API to delete an Loans Account inside Baopenbank
      parameters:
        - name: mobileNumber
          in: query
          required: true
          schema:
            type: string
            example: 0987654321
      responses:
        200:
          description: HTTP Status OK
        417:
          description: HTTP Status Exception Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"



components:
  schemas:
    Loan:
      description: Schema to hold Loan information
      type: object
      properties:
        mobile_number:
          type: string
          description: Mobile number of Customer
          example: 0987654321
        loan_number:
          type: string
          description: Loan number of Loan Account
          example: 100320245735
        loan_type:
          type: string
          description: Type of Loan Account
          example: "SAVINGS"
        total_loan:
          type: integer
          description: Total amount of Loan account
          example: 10001
        amount_paid:
          type: number
          description: Paid amount of Loan account
          example: 10000
        outstanding_amount:
          type: number
          description: Outstanding amount of Loan account
          example: 10000



    ErrorResponse:
      description: Scheme to hold errorResponse information
      type: object
      properties:
        apiPath:
          type: string
          description: API Path invoked by client
        errorCode:
          type: string
          description: Error code representing the error Code
        errorMessage:
          type: string
          description: Error message representing the error details
        timeStamp:
          type: string
          description: Timestamp representing when the request is invoked
          format: date-time


    ResponseDetails:
      description: Scheme to hole response details
      type: object
      properties:
        statusCode:
          type: string
        statusMessage:
          type: string